Coding is cool

+++ 0x00-hello_world/6-size.c
@@ -17,10 +17,10 @@ int main(void)
 	float(f);
 
 
-	printf("Size of a char: %d byte(s)\n", (int) sizeof(c));
-	printf("Size of an int: %d byte(s)\n", (int) sizeof(i));
-	printf("Size of a long int: %d byte(s)\n", (int) sizeof(li));
-	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(lli));
-	printf("Size of a float: %d byte(s)\n", (int) sizeof(f));
+	printf("Size of a char: %ld byte(s)\n", sizeof(c));
+	printf("Size of an int: %ld byte(s)\n", sizeof(i));
+	printf("Size of a long int: %ld byte(s)\n", sizeof(li));
+	printf("Size of a long long int: %ld byte(s)\n", sizeof(lli));
+	printf("Size of a float: %ld byte(s)\n", sizeof(f));
 	return (0);
 }
